requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
opentelemetry-api>=1.18.0
opentelemetry-sdk>=1.18.0
opentelemetry-exporter-otlp>=1.18.0
boto3>=1.28.0